Implement message splitting in case the payload is too large
